pyproject.toml
[project]
name = "memoriaapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Natalia",email = "lasinforosa@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "django-import-export (>=4.4.0,<5.0.0)",
    "django-filter (>=25.2,<26.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "whitenoise (>=6.11.0,<7.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
